Look in current directory for support exe instead of asking user if we should
Other than automatically checking for local copies of ffmpeg and yt-dlp, this is the
same as the previous operation. The script will check in all pre-existing PATH locations
before looking in the current directory. IT SHOULD BE NOTED: That is the CURRENT
directory, not the SCRIPT directory. If the user runs the script from some other
directory then it will look there instead. This was an issue in the old method as well.